renaming endpoints to implementations
