+++ cfme/automate/__init__.py
@@ -6,7 +6,7 @@ from widgetastic_patternfly import Accordion, Dropdown
 
 from cfme import BaseLoggedInPage
 from cfme.base import Server
-from utils.appliance.endpoints.ui import navigator, CFMENavigateStep
+from utils.appliance.implementations.ui import navigator, CFMENavigateStep
 
 
 class AutomateCustomizationView(BaseLoggedInPage):
