+++ cfme/containers/image.py
@@ -7,7 +7,7 @@ from navmazing import NavigateToSibling, NavigateToAttribute
 from cfme.common import SummaryMixin, Taggable
 from cfme.fixtures import pytest_selenium as sel
 from cfme.web_ui import toolbar as tb, CheckboxTable, paginator, match_location, InfoBlock
-from utils.appliance.endpoints.ui import CFMENavigateStep, navigator, navigate_to
+from utils.appliance.implementations.ui import CFMENavigateStep, navigator, navigate_to
 from utils.appliance import Navigatable
 
 list_tbl = CheckboxTable(table_locator="//div[@id='list_grid']//table")
