+++ cfme/tests/automate/test_provisioning_dialogs.py
@@ -6,7 +6,7 @@ from cfme import test_requirements
 from cfme.automate import provisioning_dialogs
 from cfme.web_ui import accordion
 from utils.update import update
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 
 
 @pytest.yield_fixture(scope="function")
