+++ cfme/tests/cloud/test_stack.py
@@ -3,7 +3,7 @@ import pytest
 from cfme.fixtures import pytest_selenium as sel
 from cfme.cloud.stack import Stack
 from utils import testgen
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 from cfme.configure import settings  # NOQA
 from cfme.web_ui import Table, Quadicon
