+++ cfme/tests/containers/test_views_objects.py
@@ -8,7 +8,7 @@ from cfme.containers.node import Node
 from cfme.fixtures import pytest_selenium as sel
 from cfme.web_ui import toolbar as tb
 from utils import testgen
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 from utils.version import current_version
