+++ cfme/tests/infrastructure/test_advanced_search_host.py
@@ -11,7 +11,7 @@ from cfme.web_ui import search
 from cfme.web_ui.cfme_exception import (assert_no_cfme_exception,
     is_cfme_exception, cfme_exception_text)
 from utils.version import current_version
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 
 
 @pytest.fixture(scope="module")
