+++ cfme/tests/infrastructure/test_set_default_filter.py
@@ -8,7 +8,7 @@ from cfme.infrastructure import host, datastore
 from utils.providers import setup_a_provider
 from utils import version
 from cfme.web_ui import accordion, listaccordion as list_acc
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 from cfme.infrastructure.host import Host
 from cfme.infrastructure.datastore import Datastore
