+++ cfme/tests/intelligence/reports/test_widgets.py
@@ -5,7 +5,7 @@ import pytest
 from cfme.dashboard import Widget
 from cfme.intelligence.reports.widgets import MenuWidget, ReportWidget, RSSFeedWidget, ChartWidget
 from cfme.intelligence.reports.dashboards import DefaultDashboard
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 from utils.update import update
 from cfme import test_requirements
