+++ cfme/tests/intelligence/test_dashboard.py
@@ -10,7 +10,7 @@ from cfme.dashboard import Widget
 from cfme.intelligence.reports.dashboards import Dashboard, DefaultDashboard
 from utils.blockers import BZ
 from cfme import test_requirements
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 
 
 pytestmark = [
