+++ cfme/tests/openstack/test_roles_names.py
@@ -5,7 +5,7 @@ from cfme.web_ui import Quadicon
 from cfme.infrastructure.cluster import Cluster
 
 from utils import testgen
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 
 
 pytest_generate_tests = testgen.generate(testgen.provider_by_type,
