+++ cfme/tests/test_csrf.py
@@ -4,7 +4,7 @@ import pytest
 from cfme import dashboard
 from cfme.login import LoginPage
 from cfme.base.ui import Server
-from utils.appliance.endpoints.ui import navigate_to
+from utils.appliance.implementations.ui import navigate_to
 from utils.wait import wait_for, TimedOutError
